requirements.txt
streamlit==1.28.0
pandas==2.1.1
python-dotenv==1.0.0
fpdf2==2.7.5
groq